==> R_notebooks/STEP_1_Aim_1_Data_Setup_v3.Rmd <==
---
title: An R Markdown document converted from "STEP_1_Aim_1_Data_Setup_v3.ipynb"
output: html_document
---

```{r}
# data prep #
install.packages(c('rstatix', 'openxlsx', 'reshape2', 'janitor', 'bigrquery'))
```

```{r}
library(tidyverse)
library(bigrquery)
library(rstatix)
library(openxlsx)
library(reshape2)
library(janitor)
# bq_auth(use_oob = TRUE, cache = FALSE)
```

```{r}
## JIVE Paper Information

# Male IDs
male.jive.id.raw <- openxlsx::read.xlsx("../data/aao5253_tables_s1_to_s12.xlsx", sheet = "Supplemental Table 2", colNames = F, rows = 5:10)
male.jive.id.list <- reshape2::melt(male.jive.id.raw[,-2], id = "X1") %>% rename(cluster.group = X1) %>% select(-variable) %>% filter(!is.na(value)) %>% arrange(cluster.group)

# Female IDs
female.jive.id.raw <- openxlsx::read.xlsx("../data/aao5253_tables_s1_to_s12.xlsx", sheet = "Supplemental Table 2", colNames = F, rows = 21:25)
female.jive.id.list <- reshape2::melt(female.jive.id.raw[,-2], id = "X1") %>% rename(cluster.group = X1) %>% select(-variable) %>% filter(!is.na(value)) %>% arrange(cluster.group)

# Male Gene List
male.jive.gene.raw <- openxlsx::read.xlsx("../data/aao5253_tables_s1_to_s12.xlsx", sheet = "Supplemental Table 2", colNames = F, rows = 12:16)
male.jive.gene.list <- reshape2::melt(male.jive.gene.raw[,-2], id = "X1") %>% rename(cluster.group = X1) %>% select(-variable) %>% filter(!is.na(value)) %>% arrange(cluster.group)

# Female Gene List
female.jive.gene.raw <- openxlsx::read.xlsx("../data/aao5253_tables_s1_to_s12.xlsx", sheet = "Supplemental Table 2", colNames = F, rows = 28:32)
female.jive.gene.list <- reshape2::melt(female.jive.gene.raw[,-2], id = "X1") %>% rename(cluster.group = X1) %>% select(-variable) %>% filter(!is.na(value)) %>% arrange(cluster.group)
```

```{r}
# Filtering genes to those that code proteins
gene_table <- read_table('../data/ncbi_gene_table_dec272022.tsv.gz')
prot_coding <- gene_table %>% dplyr::filter(Gene_Type == 'PROTEIN_CODING')
```

```{r}
## Array data for training

## see raw_affy_array_processing
```

```{r}
## TCGA Query
TCGA <- "
    SELECT gexp.case_id, gexp.gene_name, clin.demo__gender, clin.demo__vital_status, clin.demo__days_to_death, clin.diag__age_at_diagnosis,clin.demo__race,clin.demo__ethnicity,
    AVG( LOG10(gexp.fpkm_uq_unstranded +1 ) ) as gene_exp
    FROM gbm-gender.tcga_gbm.rnaseq_gencode36_hg38 as gexp
    JOIN isb-cgc-bq.TCGA_versioned.clinical_gdc_r31 as clin
    ON gexp.case_id = clin.submitter_id
    WHERE gexp.sample_type = 'Primary Tumor'
        AND gene_type = 'protein_coding'
        AND primary_site = 'Brain'
    GROUP BY gexp.case_id, gene_name, clin.demo__gender, clin.diag__age_at_diagnosis,demo__race,clin.demo__ethnicity, clin.demo__vital_status,clin.submitter_id, 
             clin.demo__days_to_death, clin.diag__days_to_last_follow_up,clin.exp__created_datetime
    "

TCGA2 <- "
    SELECT gexp.case_id, gexp.gene_name, clin.demo__gender, clin.demo__vital_status, clin.demo__days_to_death, clin.diag__age_at_diagnosis,clin.demo__race,clin.demo__ethnicity,
    AVG( LOG10(gexp.unstranded +1 ) ) as gene_exp
    FROM gbm-gender.tcga_gbm.rnaseq_gencode36_hg38 as gexp
    JOIN isb-cgc-bq.TCGA_versioned.clinical_gdc_r31 as clin
    ON gexp.case_id = clin.submitter_id
    WHERE gexp.sample_type = 'Primary Tumor'
        AND gene_type = 'protein_coding'
        AND primary_site = 'Brain'
    GROUP BY gexp.case_id, gene_name, clin.demo__gender, clin.diag__age_at_diagnosis,demo__race,clin.demo__ethnicity, clin.demo__vital_status,clin.submitter_id, 
             clin.demo__days_to_death, clin.diag__days_to_last_follow_up,clin.exp__created_datetime
    "

```

```{r}
tcga.import <- bq_project_query('gbm-gender', query = TCGA2) #Put data in temporary BQ table

tcga_dwnld <- bq_table_download(tcga.import) #Put data into a dataframe

write.table(tcga_dwnld, file = '../data/rnaseq_dwnld.csv', quote = F, row.names = F, sep = ',')

#tcga_dwnld <- read_csv('../data/bq-results-20240316__TCGA_RNA_seq_unstranded.csv')

head(tcga_dwnld)

rnaseq.data<-tcga_dwnld %>% 
     pivot_wider(id_cols = case_id, names_from = gene_name, values_from = gene_exp) %>% rename(sample = case_id)
```



```{r}
### Preprocessing RNA-seq data
### remove columns of all zeros
idx <- which( apply( as.matrix(rnaseq.data[,2:ncol(rnaseq.data)]), 2, sum, na.rm=TRUE) > 0 )
rnaseq.data <- rnaseq.data[, c(1, idx+1)]

### remove columns with zero variance
jdx <- which( apply(as.matrix(rnaseq.data[,2:ncol(rnaseq.data)]), 2, var, na.rm=TRUE) > 0 )
rnaseq.data <- rnaseq.data[, c(1,jdx+1)]
```


```{r}
# processing the RNA-seq data #

male_ids <- str_sub(male.jive.id.list$value, start=1, end=12)
female_ids <- str_sub(female.jive.id.list$value, start=1, end=12)

male.jive.id.list$sample = male_ids
female.jive.id.list$sample = female_ids

tgca.rnaseq.jive <- rnaseq.data[rnaseq.data$sample %in% c(male_ids,female_ids),]

rnaseq.m <- tgca.rnaseq.jive %>% inner_join(. , male.jive.id.list, by = c("sample" = "sample"))

rnaseq.f <- tgca.rnaseq.jive %>% inner_join(. , female.jive.id.list, by = c("sample" = "sample"))
```

```{r}
dim(rnaseq.m)
dim(rnaseq.f)
```

```{r}
write.csv(rnaseq.m,"../data/Males_RNASeq_Data_unstranded.csv",row.names=F)
write.csv(rnaseq.f,"../data/Females_RNASeq_Data_unstranded.csv",row.names=F)
```






```{r}
# TEMPUS DATA
TEMPUS <- "
    SELECT gexp.case_id, clin.Sex, clin.Age_at_Diagnosis, clin.survival_status, clin.Survival_days, gexp.gene_name, AVG( LOG10(gexp.fpkm_uq_unstranded +1 ) ) as gene_exp
        FROM `tempus_group_level.rnaseq_gencode36_hg38` as gexp
        JOIN `tempus_group_level.clin` as clin
        ON gexp.case_id = clin.Subject_ID
            AND gexp.gene_type = 'protein_coding'
    GROUP BY case_id, gene_name, clin.Sex, clin.Age_at_Diagnosis, clin.survival_status, clin.Survival_days"


TEMPUS2 <- "
    SELECT gexp.case_id, clin.Sex, clin.Age_at_Diagnosis, clin.survival_status, clin.Survival_days, gexp.gene_name, AVG( LOG10(gexp.unstranded +1 ) ) as gene_exp
        FROM `tempus_group_level.rnaseq_gencode36_hg38` as gexp
        JOIN `tempus_group_level.clin` as clin
        ON gexp.case_id = clin.Subject_ID
            AND gexp.gene_type = 'protein_coding'
    GROUP BY case_id, gene_name, clin.Sex, clin.Age_at_Diagnosis, clin.survival_status, clin.Survival_days"


tempus.import <- bq_project_query ('gbm-gender', query = TEMPUS2) #Put data in temporary BQ table

tempus_dwnld<- bq_table_download(tempus.import) #Put data into a dataframe

write.table(tempus_dwnld, file = '../data/tempus_dwnld_unstranded.csv', quote = F, row.names = F, sep = ',')




```


```{r}
tempus_dwnld <- read_csv('../data/tempus_dwnld.csv')

tempus_males <-tempus_dwnld %>% filter(Sex == "male") %>%
     pivot_wider(id_cols = case_id, names_from = gene_name, values_from = gene_exp) %>% rename(sample = case_id)

tempus_females <-tempus_dwnld %>% filter(Sex == "female") %>%
     pivot_wider(id_cols = case_id, names_from = gene_name, values_from = gene_exp) %>% rename(sample = case_id)

clin_male <- tempus_dwnld %>% filter(Sex == "male") %>% 
    select(case_id,survival_status,Survival_days) %>% 
    rename(sample = case_id) %>%
    unique()

clin_female <- tempus_dwnld %>% filter(Sex == "female") %>% 
    select(case_id,survival_status,Survival_days) %>% 
    rename(sample = case_id) %>%
    unique()

tempus_male2 <- inner_join(tempus_males, clin_male)
tempus_female2 <- inner_join(tempus_females, clin_female)

dim(tempus_male2)
dim(tempus_female2)
```


```{r}
#tempus.m.test$cluster.group <- NA
write.csv(tempus_male2,"../data/Tempus_Males_unstranded.csv",row.names=F)

#tempus.f.test$cluster.group <- NA
write.csv(tempus_female2,"../data/Tempus_Females_unstranded.csv",row.names=F)
```


NOW for adding CPTAC3!

```{r}
cptac_query <- "
  SELECT
    gexp.case_gdc_id, gexp.gene_name, clin.submitter_id,
    clin.demo__gender, clin.demo__vital_status, clin.demo__days_to_death,
    clin.diag__days_to_last_known_disease_status,
    clin.diag__primary_diagnosis, clin.diag__age_at_diagnosis,
    clin.demo__race, clin.demo__ethnicity,
    AVG( LOG10(gexp.unstranded+1 ) ) as gene_exp

    FROM isb-cgc-bq.CPTAC.RNAseq_hg38_gdc_current as gexp
    JOIN isb-cgc-bq.CPTAC.clinical_gdc_current as clin
    ON gexp.case_gdc_id = case_id
   
    WHERE gexp.sample_type_name = 'Primary Tumor'
        AND gene_type = 'protein_coding'
        AND gexp.primary_site = 'Brain'

    GROUP BY
    gexp.case_gdc_id, gene_name, clin.submitter_id, clin.demo__gender,
    clin.diag__age_at_diagnosis, demo__race, clin.demo__ethnicity,
    clin.demo__vital_status, clin.submitter_id, clin.diag__primary_diagnosis,
    clin.demo__days_to_death, clin.diag__days_to_last_follow_up,clin.exp__created_datetime,
    clin.lost_to_followup,clin.diag__days_to_last_known_disease_status"

cptac3.import <- bq_project_query('gbm-gender', query = cptac_query) #Put data in temporary BQ table

cptac3_dwnld <- bq_table_download(cptac3.import) #Put data into a dataframe

write.table(cptac3_dwnld, file = '../data/cptac3_rnaseq_dwnld.csv', quote = F, row.names = F, sep = ',')

cptac3.rnaseq.data <- cptac3_dwnld %>% 
     pivot_wider(id_cols = submitter_id, names_from = gene_name, values_from = gene_exp) %>% rename(sample = submitter_id)
```
```{r}
### Preprocessing RNA-seq data
### remove columns of all zeros
idx <- which( apply( as.matrix(cptac3.rnaseq.data[,2:ncol(cptac3.rnaseq.data)]), 2, sum, na.rm=TRUE) > 0 )
cptac3.rnaseq.data <- cptac3.rnaseq.data[, c(1, idx+1)]

### remove columns with zero variance
jdx <- which( apply(as.matrix(cptac3.rnaseq.data[,2:ncol(cptac3.rnaseq.data)]), 2, var, na.rm=TRUE) > 0 )
cptac3.rnaseq.data <- cptac3.rnaseq.data[, c(1,jdx+1)]
```


```{r}

cptac3_pheno_full <- unique(cptac3_dwnld[,c('case_gdc_id','submitter_id',
                                       'demo__gender','demo__vital_status',
                                       'demo__days_to_death',
                                       'diag__days_to_last_known_disease_status',
                                       'diag__primary_diagnosis',
                                       'diag__age_at_diagnosis','demo__race',
                                       'demo__ethnicity')])


cptac3_pheno <- unique(cptac3_dwnld[,c('submitter_id',
                                       'demo__gender','demo__vital_status',
                                       'diag__days_to_last_known_disease_status')])

cptac3_pheno_f <- cptac3_pheno %>% filter(demo__gender == 'female')
cptac3_pheno_m <- cptac3_pheno %>% filter(demo__gender == 'male')


cptac3.rnaseq.m <- cptac3.rnaseq.data %>% inner_join(. , cptac3_pheno_m, by = c("sample" = "submitter_id"))
cptac3.rnaseq.f <- cptac3.rnaseq.data %>% inner_join(. , cptac3_pheno_f, by = c("sample" = "submitter_id"))


```


```{r}
write.csv(cptac3.rnaseq.m,"../data/CPTAC3_Males_RNASeq_Data_unstranded.csv",row.names=F)
write.csv(cptac3.rnaseq.f,"../data/CPTAC3_Females_RNASeq_Data_unstranded.csv",row.names=F)
```
