==> R_notebooks/STEP_2_1_Data_Bundling.Rmd <==
---
title: "Data Bundling"
output: html_notebook
---

This notebook is to process the data for fast loading and use later on.

```{r}

# TCGA array data
colnames(dat_f)[12645:12647] # "Barcode"      "ClusterLabel" "Sex"       

# TCGA rna-seq data
colnames(tcga_f)[1] <- 'Barcode'
colnames(tcga_f)[19428:19429] <- c('ClusterLabel', 'SampleBarcode')  #"cluster.group" "value" 

# Validation array data sets
colnames(val_f)[11916:11919] <- c('Barcode','Sex','Survival', 'Censored') # "SampleID" "Sex"      "Survival" "Censored"

# Tempus rna-seq data
colnames(tmp_f)[19939:19940] <- c('Censored','Survival')
colnames(tmp_f)[1] <- 'Barcode'

# CPTAC3 rna-seq data
cpt_f <- cptac3.rnaseq.f
colnames(cpt_f)[1] <- 'Barcode'
colnames(cpt_f[19452:19454]) <- c('Sex','Censored','Survival')
 # "demo__gender"                            "demo__vital_status"                      "diag__days_to_last_known_disease_status"

save(dat_f, tcga_f, val_f, tmp_f, cpt_f, file='../data/F_processed_data.rda')

```


```{r}

# collecting the male data into a convient package

dat_m <- read.csv('../data/jive_training_array_data_M.csv.gz')
dat_m <- dat_m[,-1]
# Barcode ClusterLabel Sex

load('../data/joined_validation_array_data_M.rda')
#SampleID[11916]  Sex[11917]   Survival[11918] Censored[11919]
colnames(val_m)[11916:11919] <- c('Barcode', 'Sex', 'Survival', 'Censored')


rna_m <- read.csv("../data/Males_RNASeq_Data_unstranded.csv.gz")
# c(1,19428,19429 )
# sample [1] cluster.group value (barcode)
colnames(rna_m)[c(1,19428,19429)] <- c('Barcode', 'ClusterLabel', 'SampleBarcode')

tmp_m <- read.csv("../data/Tempus_Males_unstranded.csv.gz")
tmp_m[['RNASE4']]  <- NULL # has 0 var
#sample[1]   survival_status[19940]  Survival_days[19941]
colnames(tmp_m)[c(1,19939,19940)] <- c('Barcode', 'Censored', 'Survival')


cpt_m <- read.csv("../data/CPTAC3_Males_RNASeq_Data_unstranded.csv")
#sample[1] demo__gender[19452] demo__vital_status[19453] diag__days_to_last_known_disease_status[19454]
colnames(cpt_m)[c(1,19452,19453,19454)] <- c('Barcode', 'Sex', 'Censored', 'Survival')

save(cpt_m, dat_m, rna_m, tmp_m, val_m, file='../data/M_processed_data.rda')

```



SET up the gene signatures

```{r}

load('../results/results_no_array_min/females_genepairs.rda')
gp_nomin <- genepairs
load('../results/results_median_min/females_genepairs.rda')
gp1 <- genepairs
load('../results/females_genepairs_val.rda')
gp2 <- genepairs
load('../results/females_genepairs_rna.rda')
gp3 <- genepairs
load('../results/females_genepairs_refined.rda')
gp4 <- genepairs

fgenepairs <- list()
fgenepairs[['no_min']] <- gp_nomin
fgenepairs[['median_min']] <- gp1
fgenepairs[['val']] <- gp2
fgenepairs[['rna']] <- gp3
fgenepairs[['refined']] <- gp4

save(fgenepairs, file='../data/f_genepairs.rda')


load('../results/results_no_array_min/males_genepairs.rda')
gp_nomin <- genepairs
load('../results/results_median_min/males_genepairs.rda')
gp1 <- genepairs
load('../results/males_genepairs_val.rda')
gp2 <- genepairs
load('../results/males_genepairs_rna.rda')
gp3 <- genepairs

mgenepairs <- list()
mgenepairs[['no_min']] <- gp_nomin
mgenepairs[['median_min']] <- gp1
mgenepairs[['val']] <- gp2
mgenepairs[['rna']] <- gp3

save(mgenepairs, file='../data/m_genepairs.rda')

```
