==> R_notebooks/STEP_4_AIM_1_CV_Pairs.Rmd <==
---
title: An R Markdown document converted from "STEP_4_AIM_1_CV_Pairs.ipynb"
output: html_document
---

Changes:

* Subsetting to genes that are present in all data sources.. array, RNA-seq, and TEMPUS. 10363 genes.
* Subsetting the JIVE gene signatures to those present genes.

Questions: Are there missing genes, in array, that are present in RNA-seq, that would be better to have for TEMPUS?



```{r}
#devtools::install_github("gibbsdavidl/robencla", ref = 'v0.3.4')
```

```{r}
library(tidyverse)
library(robencla)
library(pheatmap)
```

```{r}

# pairlist is a list of vectors (paired terms)
# n is the number of pairs for each entry desired

shorter_and_clean <- function(pairlist, n) {
  newlist <- list()
  for (ni in names(pairlist)) {
    newlist[[ni]] <- str_replace_all(pairlist[[ni]][1:(2*n)], '\\.', '_')
  }
  return(newlist)  
}


```



```{r}

load('../data/M_processed_data.rda')
load('../results/males_genepairs.rda')

# removing signature genes that are not available in our data sets
male.jive.gene.list <- readRDS("../data/Males_jive_cluster_genes.rds") %>% dplyr::filter(value %in% colnames(dat_m))


sigs_m <- list("C1"=male.jive.gene.list$value[male.jive.gene.list$cluster.group == 'cluster1'],
               "C2"=male.jive.gene.list$value[male.jive.gene.list$cluster.group == 'cluster2'],
               "C3"=male.jive.gene.list$value[male.jive.gene.list$cluster.group == 'cluster3'],
               "C4"=male.jive.gene.list$value[male.jive.gene.list$cluster.group == 'cluster4'],
               "C5"=male.jive.gene.list$value[male.jive.gene.list$cluster.group == 'cluster5']
)

```

```{r}
table(dat_m$ClusterLabel)
```

```{r}
dim(dat_m)
print(tail(colnames(dat_m)))
```

```{r}


# our classifier object named Roberta
mod_m <- Robencla$new("mod_m")

# xgboost parameters to pass to each sub-classifier in the ensembles
params <- list(max_depth=12,   # "height" of the tree,   (xgboost parameter)
               eta=0.45,       # this is the learning rate. smaller values slow it down  (xgboost parameter)
               nrounds=24,     # number of rounds of training, lower numbers less overfitting (xgboost parameter)
               early_stopping_rounds=2,
               nthreads=8,     # parallel threads
               gamma=0.2,      # Minimum loss req'd to again partition a leaf node. higher ~ more conservative (xgboost)
               lambda=2.5,     # L2 regularization term on weights, higher ~ more conservative (xgboost parameter)
               alpha=0.25,     # L1 regularization term on weights. higher ~ more conservative (xgboost parameter)
               verbose=0,
               train_perc=0.8,
               combine_function='median',
               size=11
)

gpairs <- shorter_and_clean(genepairs, 10)

# First we use the training data
mod_m$autocv(data_frame=dat_m,
            label_name='ClusterLabel',
            sample_id = 'Barcode',
            drop_list='Sex',
            data_mode=c('pairs'), # pairs, allpairs, sigpairs, quartiles, tertiles, binarize, ranks, original #
            signatures=NULL, 
            pair_list=gpairs, # subset to these genes.
            params=params,
            cv_rounds=10
            )


```

```{r}
df.m <- mod_m$cv_results
df.m$cluster1 <- as.numeric(df.m$cluster1)
df.m$cluster2 <- as.numeric(df.m$cluster2)
df.m$cluster3 <- as.numeric(df.m$cluster3)
df.m$cluster4 <- as.numeric(df.m$cluster4)
df.m$cluster5 <- as.numeric(df.m$cluster5)
head(df.m)
df.m
```

```{r}


save_pheatmap_pdf <- function(x, filename, width=7, height=7) {
    stopifnot(!missing(x))
    stopifnot(!missing(filename))
    pdf(filename, width=width, height=height)
    grid::grid.newpage()
    grid::grid.draw(x$gtable)
    dev.off()
}

tabdf <- table(Pred=df.m$BestCalls, True=df.m$Label)

print(tabdf)

tabtbl <- as.data.frame(tabdf) %>% 
  pivot_wider(id_cols = Pred, names_from = True, values_from = Freq) %>% 
  select(-Pred) #%>%
  #scale(center = F) # Z-scores
rownames(tabtbl) <- colnames(tabtbl)

p <- pheatmap(tabtbl, cluster_rows = F, cluster_cols = F, display_numbers = T, 
              scale = 'column', fontsize_number = 12)

#           True
# Pred       cluster1 cluster2 cluster3 cluster4 cluster5
#   cluster1       32        7        1        2        1
#   cluster2        4       75        2        1        2
#   cluster3        1        0       27        0        2
#   cluster4        1        0        1       22        0
#   cluster5        0        3        4        1       31
```




```{r}
metrics <- mod_m$classification_metrics()

metrics
```

```{r}

imp_list <- mod_m$importance()

imp_list

```


And to write everything out

```{r}

save(mod_m, file='../results/models/M_TCGA_CV.rda')

write_csv(metrics, file='../results/males_TCGA_CV_metrics_step4.csv')

write.csv(as.data.frame(df.m), '../results/male_TCGA_CV.csv', sep=',', quote = F, row.names = F)

save(imp_list, file='../results/male_TCGA_CV_pairs_importance_step4.rda')

save_pheatmap_pdf(p, "../figures/males_TCGA_CV_heatmap.pdf", width=5, height=5)

```

```{r}

df1 <- data.frame(Top10=1:10, Cluster='cluster1',imp_list[['cluster1']][1:10,])
df2 <- data.frame(Top10=1:10, Cluster='cluster2',imp_list[['cluster2']][1:10,])
df3 <- data.frame(Top10=1:10, Cluster='cluster3',imp_list[['cluster3']][1:10,])
df4 <- data.frame(Top10=1:10, Cluster='cluster4',imp_list[['cluster4']][1:10,])
df5 <- data.frame(Top10=1:10, Cluster='cluster5',imp_list[['cluster5']][1:10,])

impdf <- rbind(df1[,c(1,2,4)],df2[,c(1,2,4)],df3[,c(1,2,4)],df4[,c(1,2,4)],df5[,c(1,2,4)])

colnames(impdf) <- c('Top10','Cluster','Med_Info_Gain')


ggplot(impdf, aes(x=Top10,y=Med_Info_Gain,color=Cluster)) + geom_line() + 
  xlab("Top 10 features per cluster") + ylab("Median Information Gain")
ggsave('../figures/males_TCGA_CV_feature_importance.pdf')

allimp <- rbind(df1,df2,df3,df4,df5)

write.csv(allimp, file='../results/male_importance_table.csv')

```







FEMALES





```{r}

load('../data/F_processed_data.rda')
load('../results/females_genepairs.rda')


# removing signature genes that are not available in our data sets
female.jive.gene.list <- readRDS("../data/Female_jive_cluster_genes.rds") %>% dplyr::filter(value %in% colnames(dat_f))

sigs_f <- list("C1"=female.jive.gene.list$value[female.jive.gene.list$cluster.group == 'cluster1'],
               "C2"=female.jive.gene.list$value[female.jive.gene.list$cluster.group == 'cluster2'],
               "C3"=female.jive.gene.list$value[female.jive.gene.list$cluster.group == 'cluster3'],
               "C4"=female.jive.gene.list$value[female.jive.gene.list$cluster.group == 'cluster4'],
               "C5"=female.jive.gene.list$value[female.jive.gene.list$cluster.group == 'cluster5']
)

```

```{r}
table(dat_f$ClusterLabel)
```



```{r}

# our classifier object named Roberta
buffy <- Robencla$new("fmod")

# xgboost parameters to pass to each sub-classifier in the ensembles
params <- list(max_depth=12,   # "height" of the tree, 6 is actually default.   (xgboost parameter)
               eta=0.45,        # this is the learning rate. smaller, more  conservative   (xgboost parameter)
               nrounds=24,     # number of rounds of training,  (xgboost parameter)
               early_stopping_rounds=2,
               nthreads=12,     # parallel threads
               gamma=1.25,      # Minimum loss req'd to again partition a leaf node. higher ~ more conservative (xgboost)
               lambda=2.5,     # L2 regularization term on weights, higher ~ more conservative (xgboost parameter)
               alpha=1.25,      # L1 regularization term on weights. higher ~ more conservative (xgboost parameter)
               verbose=0,
               train_perc=0.8,
               combine_function='median',
               size=11
)


glist <- shorter_and_clean(genepairs, 10)

# First we use the training data
buffy$autocv(data_frame=dat_f,
            label_name='ClusterLabel',
            sample_id = 'Barcode',
            drop_list = 'Sex',
            data_mode=c('pairs'), # pairs, allpairs, sigpairs, quartiles, tertiles, binarize, ranks, original #
            signatures=NULL,
            pair_list=glist,  # subset to these genes.
            params=params,
            cv_rounds=10
            )

```

```{r}
df.f <- buffy$cv_results
df.f$cluster1 <- as.numeric(df.f$cluster1)
df.f$cluster2 <- as.numeric(df.f$cluster2)
df.f$cluster3 <- as.numeric(df.f$cluster3)
df.f$cluster4 <- as.numeric(df.f$cluster4)
df.f$cluster5 <- as.numeric(df.f$cluster5)
head(df.f)

```

```{r}

save_pheatmap_pdf <- function(x, filename, width=7, height=7) {
    stopifnot(!missing(x))
    stopifnot(!missing(filename))
    pdf(filename, width=width, height=height)
    grid::grid.newpage()
    grid::grid.draw(x$gtable)
    dev.off()
}

tabdf <- table(Pred=df.f$BestCalls, True=df.f$Label)

print(tabdf)

tabtbl <- as.data.frame(tabdf) %>% 
  pivot_wider(id_cols = Pred, names_from = True, values_from = Freq) %>% 
  select(-Pred) #%>%
  #scale(center = F) # Z-scores
rownames(tabtbl) <- colnames(tabtbl)

p <- pheatmap(tabtbl, cluster_rows = F, cluster_cols = F, display_numbers = T, 
              scale = 'column', fontsize_number = 12)

#           True
# Pred       cluster1 cluster2 cluster3 cluster4 cluster5
#   cluster1       34        2        0        2        4
#   cluster2        2       20        1        2        1
#   cluster3        0        0       11        0        0
#   cluster4        3        1        0       14        0
#   cluster5        2        1        2        2       35

```

```{r}
metrics <- buffy$classification_metrics()

metrics
```

```{r}

imp_list <- buffy$importance()

imp_list

```


```{r}

df1 <- data.frame(Top10=1:10, Cluster='cluster1',imp_list[['cluster1']][1:10,])
df2 <- data.frame(Top10=1:10, Cluster='cluster2',imp_list[['cluster2']][1:10,])
df3 <- data.frame(Top10=1:10, Cluster='cluster3',imp_list[['cluster3']][1:10,])
df4 <- data.frame(Top10=1:10, Cluster='cluster4',imp_list[['cluster4']][1:10,])
df5 <- data.frame(Top10=1:10, Cluster='cluster5',imp_list[['cluster5']][1:10,])

impdf <- rbind(df1[,c(1,2,4)],df2[,c(1,2,4)],df3[,c(1,2,4)],df4[,c(1,2,4)],df5[,c(1,2,4)])

colnames(impdf) <- c('Top10','Cluster','Med_Info_Gain')

ggplot(impdf, aes(x=Top10,y=Med_Info_Gain,color=Cluster)) + geom_line() + 
  xlab("Top 10 features per cluster") + ylab("Median Information Gain")
ggsave("../figures/female_TCGA_CV_feature_importance.pdf")

allimp <- rbind(df1,df2,df3,df4,df5)

```



```{r}

save(buffy, file='../results/models/F_TCGA_CV_array.rda')

write_csv(metrics, file='../results/females_TCGA_CV_metrics_step4.csv')

save_pheatmap_pdf(p, "../figures/female_TCGA_CV_heatmap.pdf")

write.table(as.data.frame(df.f), '../results/female_TCGA_CV_pairs_step4.csv', sep=',', quote = F, row.names = F)

save(imp_list, file='../results/female_TCGA_CV_pairs_importance_step4.rda')

write.csv(allimp, file='../results/female_TCGA_CV_importance_table.csv')
#FIN
```
